move interfaces/ into workflow/; small import fixups elsewhere
